+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## Upcoming
+## Upcoming changes
+
+### ActionMenu
+
+- Prepending icons in the action menu have a fixed width, aligning the text between items
+- `ActionMenuLink` has the same indent as `ActionMenuItem`
 
 ### ActionMenuButton